Use double for locations

+++ app/src/main/java/com/csahmad/moodcloud/LocationTools.java
@@ -1,6 +1,6 @@
 package com.csahmad.moodcloud;
 
-/** Tools for locations in int[]{latitude, longitude, altitude} form */
+/** Tools for locations in double[]{latitude, longitude, altitude} form */
 public class LocationTools {
 
     // TODO: 2017-02-26 Fill out (currently just returns 0)
@@ -11,7 +11,7 @@ public class LocationTools {
      * @param location2 a location in the form {latitude, longitude, altitude}
      * @return the distance of the given locations in km
      */
-    public static int distance(int[] location1, int[] location2) {
+    public static double distance(double[] location1, double[] location2) {
 
         return 0;
     }
