+++ app/src/main/java/com/csahmad/moodcloud/Post.java
@@ -13,14 +13,14 @@ public class Post {
     private Profile poster;
 
     /** The location of the Post in the form {latitude, longitude, altitude} */
-    private int[] location;
+    private double[] location;
 
     /** This Post's unique ID (creating IDs handled by Jest). */
     @JestId
     private String id;
 
     public Post(String text, String mood, String triggerText, String triggerImage, String context,
-                Profile poster, int[] location) {
+                Profile poster, double[] location) {
 
         this.text = text;
         this.mood = mood;
@@ -99,12 +99,12 @@ public class Post {
         this.poster = poster;
     }
 
-    public int[] getLocation() {
+    public double[] getLocation() {
 
         return this.location;
     }
 
-    public void setLocation(int[] location) {
+    public void setLocation(double[] location) {
 
         this.location = location;
     }
